remove cycle between openfeature components
