+++ ddtrace/internal/openfeature/_provider.py
@@ -23,6 +23,8 @@ from openfeature.provider import ProviderStatus
 from ddtrace.internal.logger import get_logger
 from ddtrace.internal.native._native import ffe
 from ddtrace.internal.openfeature._config import _get_ffe_config
+from ddtrace.internal.openfeature._config import _register_provider
+from ddtrace.internal.openfeature._config import _unregister_provider
 from ddtrace.internal.openfeature._exposure import build_exposure_event
 from ddtrace.internal.openfeature._flag_eval_evp_hook import FlagEvalEVPHook
 from ddtrace.internal.openfeature._flageval_metrics import METADATA_ALLOCATION_KEY
@@ -623,25 +625,3 @@ class DataDogProvider(AbstractProvider):
         """
         self._exposure_cache.clear()
         logger.debug("Exposure cache cleared")
-
-
-# Module-level registry for active provider instances
-_provider_instances: list[DataDogProvider] = []
-
-
-def _register_provider(provider: DataDogProvider) -> None:
-    """Register a provider instance for configuration callbacks."""
-    if provider not in _provider_instances:
-        _provider_instances.append(provider)
-
-
-def _unregister_provider(provider: DataDogProvider) -> None:
-    """Unregister a provider instance."""
-    if provider in _provider_instances:
-        _provider_instances.remove(provider)
-
-
-def _notify_providers_config_received() -> None:
-    """Notify all registered providers that configuration was received."""
-    for provider in _provider_instances:
-        provider.on_configuration_received()
